[fixedTools] make sure fixedToFloat function returns float on both python 2 and 3

+++ Lib/fontTools/misc/fixedTools.py
@@ -26,7 +26,7 @@ def fixedToFloat(value, precisionBits):
 	hi = value + eps
 	# If the range of valid choices spans an integer, return the integer.
 	if int(lo) != int(hi):
-		return round(value)
+		return float(round(value))
 	fmt = "%.8f"
 	lo = fmt % lo
 	hi = fmt % hi
